+++ app/Http/Controllers/ProductController.php
@@ -15,6 +15,9 @@ class ProductController extends Controller
      */
     public function index()
     {
+        if (!auth()->user()->can('view products')) {
+            abort(403, 'Unauthorized action. You do not have the required permission.');
+        }
         return view('products.view');
     }
 
@@ -23,6 +26,9 @@ class ProductController extends Controller
      */
     public function create()
     {
+        if (!auth()->user()->can('create products')) {
+            abort(403, 'Unauthorized action. You do not have the required permission.');
+        }
         return view('products.create');
     }
 
@@ -38,6 +44,9 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
+        if (!auth()->user()->can('create products')) {
+            return response()->json(['data' => 'Unauthorized action. You do not have the required permission.'], 403);
+        }
         $this->validateCreateProduct($request->all())->validate();
         $product = new Product();
         $product->name = $request->name;
@@ -53,6 +62,9 @@ class ProductController extends Controller
      */
     public function show()
     {
+        if (!auth()->user()->can('view products')) {
+            return response()->json(['data' => 'Unauthorized action. You do not have the required permission.'], 403);
+        }
         return response()->json(['data' => Product::all()]);
     }
 
@@ -61,6 +73,9 @@ class ProductController extends Controller
      */
     public function edit(String $id)
     {
+        if (!auth()->user()->can('edit products')) {
+            abort(403, 'Unauthorized action. You do not have the required permission.');
+        }
         $product = Product::where('id', $id)->first()->toArray();
         return view('products.edit', compact('product'));
     }
@@ -77,6 +92,9 @@ class ProductController extends Controller
      */
     public function update(Request $request, String $id)
     {
+        if (!auth()->user()->can('edit products')) {
+            return response()->json(['data' => 'Unauthorized action. You do not have the required permission.'], 403);
+        }
         $this->validateEditProduct($request->all())->validate();
         $product = Product::where('id', $id)->first();
         $product->name = $request->name;
@@ -92,6 +110,9 @@ class ProductController extends Controller
      */
     public function destroy(Product $product, String $id)
     {
+        if (!auth()->user()->can('delete products')) {
+            return response()->json(['data' => 'Unauthorized action. You do not have the required permission.'], 403);
+        }
         $product = Product::where('id', $id)->first();
         if($product->delete()){
             return response()->json(['success' => 'User deleted!'], 200);
